Added the result count to the header of the search and paging results. Ordered all results to make consistent.  Added optional api_key that can be embedded in C64 PRG to pass in. Fixed a bug in ViceStorageAdapter that would fail when C64 filename had invalid filesystem characters.

+++ VDRIVE_Contracts/Structures/Http/HttpSearchResponse.cs
@@ -8,11 +8,11 @@ namespace VDRIVE_Contracts.Structures.Http
     public struct HttpSearchResponse
     {
         public ushort SessionId;           // Session ID (lo/hi bytes)
-        public byte ResultCount;           // Number of results found (0 if none)
-                                          // Text payload follows this header
+        public ushort ResultCount;         // Total result count (lo/hi bytes)
+                                           // Text payload follows this header
 
         // Helper method to create response with session
-        public static HttpSearchResponse Create(ushort sessionId, byte resultCount)
+        public static HttpSearchResponse Create(ushort sessionId, ushort resultCount)
         {
             return new HttpSearchResponse
             {
